refactor tables components to use common naming scheme
