+++ web-ui/src/components/table/table-denied/table-denied.less
