+++ web-ui/src/components/table/table-failures/table-failures.less
