+++ web-ui/src/components/table/table-failures/table-failures.tsx
@@ -1,10 +1,10 @@
-import { Typography, Table, Image } from 'antd';
+import { Table, Image, Typography } from 'antd';
 import { notification } from 'antd';
 import { useEffect, useState } from 'react';
 import { useMutation } from '@tanstack/react-query';
 import apiClient, { ServerErrorResponse, ServerResponse } from '@/http';
 import { Lecture } from '@/types/lecture';
-import styles from './queue-table.less';
+import styles from './table-failures.less';
 import { ColumnsType } from 'antd/es/table';
 import svFlag from '@/assets/flag-sv.svg';
 import enFlag from '@/assets/flag-en.svg';
@@ -12,13 +12,13 @@ import kthLogo from '@/assets/kth.svg';
 import youtubeLogo from '@/assets/youtube.svg';
 import {
   emitEvent,
-  CATEGORY_QUEUE_TABLE,
+  CATEGORY_FAILURE_TABLE,
   EVENT_ERROR_RESPONSE,
 } from '@/matomo';
 
-const UPDATE_INTERVAL = 10000;
-
 const { Link } = Typography;
+
+const UPDATE_INTERVAL = 5000;
 
 const columns: ColumnsType<Lecture> = [
   {
@@ -47,6 +47,17 @@ const columns: ColumnsType<Lecture> = [
       <>
         <Link href={`/lectures/${combined_public_id_and_lang}/watch`}>
           View Progress
+        </Link>
+      </>
+    ),
+  },
+  {
+    title: 'Content Link',
+    dataIndex: 'content_link',
+    render: (content_link: string) => (
+      <>
+        <Link href={content_link} target="_blank">
+          {content_link}
         </Link>
       </>
     ),
@@ -86,28 +97,6 @@ const columns: ColumnsType<Lecture> = [
     },
   },
   {
-    title: 'State',
-    dataIndex: 'state',
-    render: (val: string) => {
-      val = val.replaceAll('_', ' ');
-
-      const titleCase = val
-        .toLowerCase()
-        .replace(/(^|\s)\S/g, (match) => match.toUpperCase());
-      return <>{titleCase}</>;
-    },
-  },
-  {
-    title: 'Frozen',
-    dataIndex: 'frozen',
-    render: (isFrozen: boolean) => {
-      if (isFrozen) {
-        return <>Yes</>;
-      }
-      return <>No</>;
-    },
-  },
-  {
     title: 'Progress',
     dataIndex: 'overall_progress',
     render: (val: string) => <>{val}%</>,
@@ -118,13 +107,13 @@ interface LectureResponse extends ServerResponse {
   data: Lecture[];
 }
 
-export default function QueueTable() {
+export default function TableFailures() {
   const [lectures, setLectures] = useState<Lecture[]>([]);
   const [notificationApi, contextHolder] = notification.useNotification();
 
   const { mutate: fetchLectures } = useMutation(
     async () => {
-      return await apiClient.get(`/lectures?summary=true&only_unfinished=true`);
+      return await apiClient.get(`/lectures?summary=true&only_failed=true`);
     },
     {
       onSuccess: (res: LectureResponse) => {
@@ -145,7 +134,11 @@ export default function QueueTable() {
           message: 'Failed to get lectures',
           description: err.response.data.detail,
         });
-        emitEvent(CATEGORY_QUEUE_TABLE, EVENT_ERROR_RESPONSE, 'fetchLectures');
+        emitEvent(
+          CATEGORY_FAILURE_TABLE,
+          EVENT_ERROR_RESPONSE,
+          'fetchLectures'
+        );
       },
     }
   );
