+++ web-ui/src/components/table/table-queue/table-queue.less
