+++ web-ui/src/components/table/table-queue/table-queue.tsx
@@ -1,10 +1,10 @@
-import { Table, Image, Typography } from 'antd';
+import { Typography, Table, Image } from 'antd';
 import { notification } from 'antd';
 import { useEffect, useState } from 'react';
 import { useMutation } from '@tanstack/react-query';
 import apiClient, { ServerErrorResponse, ServerResponse } from '@/http';
 import { Lecture } from '@/types/lecture';
-import styles from './denied-table.less';
+import styles from './table-queue.less';
 import { ColumnsType } from 'antd/es/table';
 import svFlag from '@/assets/flag-sv.svg';
 import enFlag from '@/assets/flag-en.svg';
@@ -12,13 +12,13 @@ import kthLogo from '@/assets/kth.svg';
 import youtubeLogo from '@/assets/youtube.svg';
 import {
   emitEvent,
-  CATEGORY_DENIED_TABLE,
+  CATEGORY_QUEUE_TABLE,
   EVENT_ERROR_RESPONSE,
 } from '@/matomo';
 
-const { Link } = Typography;
+const UPDATE_INTERVAL = 10000;
 
-const UPDATE_INTERVAL = 5000;
+const { Link } = Typography;
 
 const columns: ColumnsType<Lecture> = [
   {
@@ -41,12 +41,12 @@ const columns: ColumnsType<Lecture> = [
     dataIndex: 'title',
   },
   {
-    title: 'Content Link',
-    dataIndex: 'content_link',
-    render: (content_link: string) => (
+    title: 'Analysis',
+    dataIndex: 'combined_public_id_and_lang',
+    render: (combined_public_id_and_lang: string) => (
       <>
-        <Link href={content_link} target="_blank">
-          {content_link}
+        <Link href={`/lectures/${combined_public_id_and_lang}/watch`}>
+          View Progress
         </Link>
       </>
     ),
@@ -97,19 +97,34 @@ const columns: ColumnsType<Lecture> = [
       return <>{titleCase}</>;
     },
   },
+  {
+    title: 'Frozen',
+    dataIndex: 'frozen',
+    render: (isFrozen: boolean) => {
+      if (isFrozen) {
+        return <>Yes</>;
+      }
+      return <>No</>;
+    },
+  },
+  {
+    title: 'Progress',
+    dataIndex: 'overall_progress',
+    render: (val: string) => <>{val}%</>,
+  },
 ];
 
 interface LectureResponse extends ServerResponse {
   data: Lecture[];
 }
 
-export default function DeniedTable() {
+export default function TableQueue() {
   const [lectures, setLectures] = useState<Lecture[]>([]);
   const [notificationApi, contextHolder] = notification.useNotification();
 
   const { mutate: fetchLectures } = useMutation(
     async () => {
-      return await apiClient.get(`/lectures?summary=true&only_denied=true`);
+      return await apiClient.get(`/lectures?summary=true&only_unfinished=true`);
     },
     {
       onSuccess: (res: LectureResponse) => {
@@ -130,7 +145,7 @@ export default function DeniedTable() {
           message: 'Failed to get lectures',
           description: err.response.data.detail,
         });
-        emitEvent(CATEGORY_DENIED_TABLE, EVENT_ERROR_RESPONSE, 'fetchLectures');
+        emitEvent(CATEGORY_QUEUE_TABLE, EVENT_ERROR_RESPONSE, 'fetchLectures');
       },
     }
   );
