+++ web-ui/src/pages/info/queue.tsx
@@ -1,8 +1,8 @@
 import Frame from '@/components/page/frame/frame';
-import QueueTable from '@/components/tables/queue-table/queue-table';
 import { registerPageLoad } from '@/matomo';
 import { useEffect } from 'react';
 import { Typography } from 'antd';
+import TableQueue from '@/components/table/table-queue/table-queue';
 
 const { Title } = Typography;
 
@@ -17,7 +17,7 @@ export default function QueuePage() {
       <Frame>
         <>
           <Title level={3}>Current queue of videos for kthGPT to watch</Title>
-          <QueueTable />
+          <TableQueue />
         </>
       </Frame>
     </>
